[IMP] mis_builder: Not display comapny name with only one effective company
The company name is displayed on the side of the account when a MIS report is marked as multi-company but has only one effective company. This is not necessary because there is only one effective company and it is not necessary to know which company the account belongs to.

MT-10671